% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeMFAseedSolved.R
\name{imputeMFAseedSolved}
\alias{imputeMFAseedSolved}
\title{Impute MFA with seed use problems solved}
\usage{
imputeMFAseedSolved(
  X,
  group,
  ncp = 2,
  type = rep("s", length(group)),
  method = c("Regularized", "EM"),
  row.w = NULL,
  coeff.ridge = 1,
  threshold = 1e-06,
  ind.sup = NULL,
  num.group.sup = NULL,
  seed = NULL,
  maxiter = 1000,
  ...
)
}
\arguments{
\item{X}{A data frame with groups of continuous or categorical variables containing missing values.}

\item{group}{A vector indicating the number of variables in each group.}

\item{ncp}{An integer corresponding to the number of components used to predict the missing entries.}

\item{type}{A vector indicating the type of variables in each group; three possibilities:
\itemize{
  \item "c" for continuous variables (the variables are centered).
  \item "s" for continuous variables (the variables are scaled to unit variance).
  \item "n" for categorical variables.
}}

\item{method}{A string indicating the imputation method; "Regularized" by default or "EM".}

\item{row.w}{A numeric vector of row weights (by default, a vector of 1 for uniform row weights).}

\item{coeff.ridge}{A numeric value; by default 1 to perform the regularized `imputeMFA` algorithm.
This is only used when `method="Regularized"`. Lower values reduce regularization, and higher values increase regularization.}

\item{threshold}{A numeric value for the threshold used to assess convergence.}

\item{ind.sup}{A vector indicating the indexes of the supplementary individuals (those not used in the imputation).}

\item{num.group.sup}{A vector indicating the group of variables that are supplementary (those not used in the imputation).}

\item{seed}{An integer value for the random seed. If `NULL` (default), missing values are initially imputed by the mean of each continuous variable and by the category proportion for categorical variables. A specific integer seed leads to random initialization.}

\item{maxiter}{An integer specifying the maximum number of iterations for the algorithm.}
}
\value{
A data frame with the imputed values.
}
\description{
This function is a version of missMDA's imputeMFA that does not crash when setting seeds.
}
